feat: update backward component
